descriptors: move derived keys into their own submodule
